+++ packages/article/__tests__/web/article.web.test.js
@@ -2,6 +2,7 @@
 import "react-native";
 import React from "react";
 import renderer from "react-test-renderer";
+import "jest-styled-components";
 import Article from "../../article";
 import shared from "../shared";