+++ packages/article/styles/meta/responsive.js
@@ -1,31 +1,26 @@
 import { Text, View } from "react-native";
-import styled from "styled-components";
-import config from "../responsive-config";
+import withResponsiveStyles from "@times-components/responsive-styles";
 
-export const MetaTextElement = styled(Text)`
-  padding: 12px 0 9px;
-
-  @media (min-width: ${config.mediumBreakpoint}) {
+export const MetaTextElement = withResponsiveStyles(Text, {
+  base: () => "padding: 12px 0 9px;",
+  mediumUp: () => `
     padding-top: 4px;
     padding-bottom: 4px;
-  }
-
-  @media (min-width: ${config.wideBreakpoint}) {
+  `,
+  wideUp: () => `
     line-height: 18px;
     padding-bottom: 25px;
-  }
-`;
+  `
+});
 
-export const Meta = styled(View)`
-  margin-left: 10px;
-  margin-right: 10px;
-
-  @media (min-width: ${config.mediumBreakpoint}) {
+export const Meta = withResponsiveStyles(View, {
+  base: () => `
+    margin-left: 10px;
+    margin-right: 10px;
+  `,
+  mediumUp: () => `
     margin-left: 0;
     margin-right: 0;
-  }
-
-  @media (min-width: ${config.wideBreakpoint}) {
-    padding-top: 0px;
-  }
-`;
+  `,
+  wideUp: () => "padding-top: 0px;"
+});