+++ packages/author-profile/author-profile-content.js
@@ -107,10 +107,13 @@ class AuthorProfileContent extends React.Component {
       });
     };
 
-    const paginationComponent = (hideResults = false) => (
+    const paginationComponent = (
+      { hideResults = false, hideTopKeyline = false } = {}
+    ) => (
       <AuthorProfilePagination
         count={count}
         hideResults={hideResults}
+        hideTopKeyline={hideTopKeyline}
         onNext={(...args) => {
           onNext(...args);
           scrollToTopNextFrame();
@@ -182,10 +185,10 @@ class AuthorProfileContent extends React.Component {
               twitter={twitter}
               onTwitterLinkPress={onTwitterLinkPress}
             />
-            {paginationComponent()}
+            {paginationComponent({ hideResults: false, hideTopKeyline: true })}
           </View>
         }
-        ListFooterComponent={paginationComponent(true)}
+        ListFooterComponent={paginationComponent({ hideResults: true })}
         ItemSeparatorComponent={() => (
           <View style={styles.padding}>
             <AuthorProfileItemSeparator />