+++ packages/article-summary/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.13.12"></a>
+## [0.13.12](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.13.11...@times-components/article-summary@0.13.12) (2018-01-11)
+
+
+### Bug Fixes
+
+* replat 1208 Article Summary margins ([#514](https://github.com/newsuk/times-components/issues/514)) ([f946d16](https://github.com/newsuk/times-components/commit/f946d16))
+
+
+
+
+<a name="0.13.11"></a>
+## [0.13.11](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.13.9...@times-components/article-summary@0.13.11) (2018-01-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-summary
+
+<a name="0.13.10"></a>
+## [0.13.10](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.13.9...@times-components/article-summary@0.13.10) (2018-01-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-summary
+
+<a name="0.13.9"></a>
+## [0.13.9](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.13.8...@times-components/article-summary@0.13.9) (2018-01-09)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-summary
+
+<a name="0.13.8"></a>
+## [0.13.8](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.13.6...@times-components/article-summary@0.13.8) (2018-01-09)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-summary
+
+<a name="0.13.7"></a>
+## [0.13.7](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.13.6...@times-components/article-summary@0.13.7) (2018-01-09)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-summary
+
 <a name="0.13.6"></a>
 ## [0.13.6](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.13.5...@times-components/article-summary@0.13.6) (2018-01-08)