+++ packages/pagination/CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.12.23"></a>
+## [0.12.23](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.12.22...@times-components/pagination@0.12.23) (2018-01-13)
+
+
+
+
+**Note:** Version bump only for package @times-components/pagination
+
+<a name="0.12.22"></a>
+## [0.12.22](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.12.21...@times-components/pagination@0.12.22) (2018-01-12)
+
+
+
+
+**Note:** Version bump only for package @times-components/pagination
+
+<a name="0.12.21"></a>
+## [0.12.21](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.12.20...@times-components/pagination@0.12.21) (2018-01-11)
+
+
+
+
+**Note:** Version bump only for package @times-components/pagination
+
+<a name="0.12.20"></a>
+## [0.12.20](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.12.19...@times-components/pagination@0.12.20) (2018-01-10)
+
+
+### Bug Fixes
+
+* apply the design review ([#512](https://github.com/newsuk/times-components/issues/512)) ([70561dd](https://github.com/newsuk/times-components/commit/70561dd))
+
+
+
+
+<a name="0.12.19"></a>
+## [0.12.19](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.12.18...@times-components/pagination@0.12.19) (2018-01-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/pagination
+
+<a name="0.12.18"></a>
+## [0.12.18](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.12.16...@times-components/pagination@0.12.18) (2018-01-09)
+
+
+
+
+**Note:** Version bump only for package @times-components/pagination
+
+<a name="0.12.17"></a>
+## [0.12.17](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.12.16...@times-components/pagination@0.12.17) (2018-01-09)
+
+
+
+
+**Note:** Version bump only for package @times-components/pagination
+
 <a name="0.12.16"></a>
 ## [0.12.16](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.12.15...@times-components/pagination@0.12.16) (2018-01-08)